idp/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
rq==1.15.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0
kubernetes==28.1.0
PyYAML==6.0.1
gitpython==3.1.40
python-terraform==0.19.0
